Refactor same-net rail alignment helpers
